add Tooltip to cells with truncate transformation
